🐛 fix: 혈당 자료형을 Double -> Integer로 수정

+++ lifeFit/src/main/java/com/kh/lifeFit/domain/healthData/HealthData.java
@@ -38,7 +38,7 @@ public class HealthData {
     private Gender userGender; // FEMALE, MALE
 
     private Double bmi;
-    private Double bloodSugar;
+    private Integer bloodSugar;
     private Integer systolic; // 수축기
     private Integer diastolic; // 이완기
